+++ cine_parameters.h
@@ -12,8 +12,8 @@ public:
     const int   nrows = 10,
     const int   prey_pop = 25,
     const int   predator_pop = 25,
-    const float prob_mutation_to_0 = 0.05,
-    const float prob_mutation_to_rd = 0.025,
+    const double prob_mutation_to_0 = 0.05,
+    const double prob_mutation_to_rd = 0.025,
     const int   timesteps = 10
     );
 
@@ -22,8 +22,8 @@ public:
     int get_nrows() const noexcept;
     int get_prey_pop() const noexcept;
     int get_predator_pop() const noexcept;
-    float get_prob_mutation_to_0() const noexcept;
-    float get_prob_mutation_to_rd() const noexcept;
+    double get_prob_mutation_to_0() const noexcept;
+    double get_prob_mutation_to_rd() const noexcept;
     int get_timesteps() const noexcept;
 
 
@@ -37,8 +37,8 @@ private:
      int   m_nrows = 10;
      int   m_prey_pop = 25;
      int   m_predator_pop = 25;
-     float m_prob_mutation_to_0 = 0.05;
-     float m_prob_mutation_to_rd = 0.025;
+     double m_prob_mutation_to_0 = 0.05;
+     double m_prob_mutation_to_rd = 0.025;
      int   m_timesteps = 10;
